Move callout show pages under events/callouts via a dedicated controller
The magic-callout show pages don't belong mixed into the registration
views. Extract an Events::CalloutsController (slug-authorized, public,
like the ticket/invoice pages) and move its views to
app/views/events/callouts/, where they resolve by convention. Routes keep
the same registration_* path helpers, now pointing at events/callouts.
RegistrationsController slims back to its own actions.
