+++ app/controllers/events/registrations_controller.rb
@@ -3,19 +3,7 @@ module Events
     before_action :authenticate_user!, only: [ :create, :destroy ]
     before_action :set_event, only: [ :create, :destroy ]
     before_action :set_registrant, only: [ :create, :destroy ]
-    before_action :set_event_registration, only: [ :show, :invoice, :scholarship, :faq,
-      :payment, :certificate, :ce, :forms, :handouts, :portal, :videoconference,
-      :resend_confirmation, :cancel, :reactivate, :pay ]
-
-    # Hidden Resource (by title) backing the handout links on the handouts page,
-    # in display order. Missing ones (e.g. not seeded in an environment) are
-    # silently skipped.
-    HANDOUT_RESOURCE_TITLES = [
-      "2-Day AWBW Facilitator Training Worksheets & Handouts",
-      "AWBW Training Workshop Worksheets",
-      "AHA Moments",
-      "Inviting and Responding to Participants' Sharing"
-    ].freeze
+    before_action :set_event_registration, only: [ :show, :invoice, :resend_confirmation, :cancel, :reactivate, :pay ]
 
     def show
       authorize! @event_registration, to: :show_public?
@@ -38,89 +26,6 @@ module Events
       authorize! @event_registration, to: :show_public?
       @event = @event_registration.event
       @invoice = EventInvoice.from_registration(@event_registration)
-    end
-
-    # Public status page for the registrant's scholarship, linked from the magic
-    # ticket callout. Shows the award (amount, funder, criteria, tasks) once a
-    # scholarship exists, or a pending state while it is only requested. When the
-    # registrant has neither requested nor received one there is nothing to show.
-    def scholarship
-      authorize! @event_registration, to: :show_public?
-
-      unless @event_registration.scholarship_requested? || @event_registration.scholarship?
-        redirect_to registration_ticket_path(@event_registration.slug)
-        return
-      end
-
-      @event = @event_registration.event
-      @scholarship = @event_registration.scholarships.first
-    end
-
-    # Public FAQ page for the training, linked from the always-present "Frequently
-    # asked questions" magic callout.
-    def faq
-      authorize! @event_registration, to: :show_public?
-      @event = @event_registration.event
-    end
-
-    # Payment page: the full allocation ledger with the running balance due.
-    def payment
-      authorize! @event_registration, to: :show_public?
-      @event = @event_registration.event
-      @allocations = @event_registration.allocations.includes(:source).order(:created_at)
-    end
-
-    # Certificate of completion, rendered like the invoice. Only reachable once
-    # the certificate is unlocked.
-    def certificate
-      authorize! @event_registration, to: :show_public?
-
-      unless @event_registration.certificate_available?
-        redirect_to registration_ticket_path(@event_registration.slug)
-        return
-      end
-
-      @event = @event_registration.event
-    end
-
-    # CE hours status: requested hours, amount owed, and license number.
-    def ce
-      authorize! @event_registration, to: :show_public?
-      @event = @event_registration.event
-    end
-
-    # Forms page: callout-card links to the W-9 (when requested), invoice (when
-    # requested), and the letter to supervisors resource — each opens in a new tab.
-    def forms
-      authorize! @event_registration, to: :show_public?
-      @event = @event_registration.event
-      @form_cards = build_form_cards
-    end
-
-    # Handouts page: callout-card links to the training worksheet/handout
-    # resources, in display order, each opening in a new tab.
-    def handouts
-      authorize! @event_registration, to: :show_public?
-      @event = @event_registration.event
-      by_title = Resource.where(title: HANDOUT_RESOURCE_TITLES).index_by(&:title)
-      @handout_cards = HANDOUT_RESOURCE_TITLES.filter_map do |title|
-        resource = by_title[title]
-        next unless resource
-        resource_card(icon: "fa-solid fa-file-pdf", title: resource.title,
-                      subtitle: "Open this training resource", href: resource_path(resource))
-      end
-    end
-
-    # Facilitator Portal access info, with a link to the home screen.
-    def portal
-      authorize! @event_registration, to: :show_public?
-      @event = @event_registration.event
-    end
-
-    # Videoconference page: the join link and add-to-calendar options.
-    def videoconference
-      authorize! @event_registration, to: :show_public?
-      @event = @event_registration.event.decorate
     end
 
     def resend_confirmation
@@ -236,35 +141,6 @@ module Events
     end
 
     private
-
-    # Builds the callout-card links shown on the forms page. W-9 and invoice only
-    # appear when the registrant requested them; the letter to supervisors is a
-    # shared resource shown whenever it has been seeded.
-    def build_form_cards
-      cards = []
-      if @event_registration.w9_requested?
-        cards << resource_card(icon: "fa-solid fa-file-pdf", title: "Download W-9",
-                               subtitle: "AWBW's W-9 tax form for your records",
-                               href: "/documents/awbw-w9.pdf", trailing_icon: "fa-solid fa-download")
-      end
-      if @event_registration.invoice_requested?
-        cards << resource_card(icon: "fa-solid fa-file-invoice-dollar", title: "View invoice",
-                               subtitle: "Itemized invoice for this registration",
-                               href: registration_invoice_path(@event_registration.slug))
-      end
-      letter = Resource.find_by(title: "Letter to Supervisors")
-      if letter
-        cards << resource_card(icon: "fa-solid fa-file-arrow-down", title: "Letter to supervisors",
-                               subtitle: "Share to request release time", href: resource_path(letter))
-      end
-      cards
-    end
-
-    # A blue, new-tab callout card linking to an external/resource document.
-    def resource_card(icon:, title:, subtitle:, href:, trailing_icon: "fa-solid fa-arrow-up-right-from-square")
-      MagicTicketCallouts::Card.new(icon_class: icon, color: "blue", title: title, subtitle: subtitle,
-                                    href: href, target: "_blank", trailing_icon: trailing_icon)
-    end
 
     def send_registration_notifications(event_registration)
       registrant_email = event_registration.registrant.preferred_email
