+++ app/views/events/callouts/_callout_page.html.erb
