+++ app/views/events/callouts/certificate.html.erb
@@ -1,6 +1,6 @@
 <% content_for(:page_bg_class, "public") %>
 <% content_for(:page_title, "Certificate of completion — #{@event.title}") %>
-<%= render layout: "events/registrations/callout_page", locals: { title: "Certificate of completion" } do %>
+<%= render layout: "events/callouts/callout_page", locals: { title: "Certificate of completion" } do %>
   <div class="rounded-xl border-2 border-purple-200 bg-purple-50/40 px-6 py-10 text-center space-y-4">
     <p class="text-xs font-semibold uppercase tracking-[0.2em] text-purple-700">Certificate of completion</p>
     <p class="text-sm text-gray-500">This certifies that</p>
