+++ app/views/events/callouts/portal.html.erb
@@ -1,6 +1,6 @@
 <% content_for(:page_bg_class, "public") %>
 <% content_for(:page_title, "Facilitator Portal access — #{@event.title}") %>
-<%= render layout: "events/registrations/callout_page", locals: { title: "Facilitator Portal access" } do %>
+<%= render layout: "events/callouts/callout_page", locals: { title: "Facilitator Portal access" } do %>
   <div class="space-y-4 text-sm text-gray-700 leading-relaxed">
     <p>Access to the Facilitator Portal is provided once you've completed both training days and your training fee is paid.</p>
     <p>You're all set — watch for your portal invitation. Once you have access, you can sign in from the home screen.</p>
