+++ app/views/events/callouts/scholarship.html.erb
