+++ spec/views/page_bg_class_alignment_spec.rb
@@ -194,15 +194,15 @@ RSpec.describe "page_bg_class alignment with policies" do
     "app/views/events/public_registrations/show.html.erb" => "public",
     "app/views/events/registrations/show.html.erb"        => "public",
     "app/views/events/registrations/invoice.html.erb"     => "public",
-    "app/views/events/registrations/scholarship.html.erb" => "public",
-    "app/views/events/registrations/faq.html.erb"         => "public",
-    "app/views/events/registrations/payment.html.erb"     => "public",
-    "app/views/events/registrations/certificate.html.erb" => "public",
-    "app/views/events/registrations/ce.html.erb"          => "public",
-    "app/views/events/registrations/forms.html.erb"       => "public",
-    "app/views/events/registrations/handouts.html.erb"    => "public",
-    "app/views/events/registrations/portal.html.erb"      => "public",
-    "app/views/events/registrations/videoconference.html.erb" => "public",
+    "app/views/events/callouts/scholarship.html.erb"      => "public",
+    "app/views/events/callouts/faq.html.erb"              => "public",
+    "app/views/events/callouts/payment.html.erb"          => "public",
+    "app/views/events/callouts/certificate.html.erb"      => "public",
+    "app/views/events/callouts/ce.html.erb"               => "public",
+    "app/views/events/callouts/forms.html.erb"            => "public",
+    "app/views/events/callouts/handouts.html.erb"         => "public",
+    "app/views/events/callouts/portal.html.erb"           => "public",
+    "app/views/events/callouts/videoconference.html.erb"  => "public",
     "app/views/events/details.html.erb"                   => "public",
     "app/views/events/ce_hours.html.erb"                  => "public",
     "app/views/registration_ticket_callouts/show.html.erb" => "public",
